Save totalAmount in store for applicationFormTable
This change induced a modification in how the totalAmounts are
calculated (entry and table). We used to have a computed property to
handle it, but we faced an unexplicated bug where when using the result
of that computed property to save the totalAmount to the store, the
computed property wasn't triggered anymore and hence not updating in the frontend.
This is why we now calculate it manually when leaving the field.