Bugfix: avoid unwanted component re-render
Problem: Suppose these are the fields in the form
```
[ input1:state0, input2:state1]
```
Suppose `input1:state0` gets changed to `input1:state1` and thereby
triggering a dynamic subform with fields `input1.1:state0`.
The current replacing mechanism, when insering the new fields ended in
the following state
```
[ input1:state1, input1.1:state0, input2:state0]
```
This due to a wrong index replacement, destroying component `input2`
and triggering all the cleanup operations in the store.

So this is fixed and should be end up in:
```
[ input1:state1, input1.1:state0, input2:state1]
```

Note: I am unsure why replace was used, instead of just creating a new
array. It didn't seem to have any side effects and simplifies the
code.