Make sure that the accordion subtitles are shown
The appuniversum update included a change which deprecated some of the AuAccordion arguments and replaced them with a shorter version. The fallback behavior contains a typo which means the subtitle was no longer being shown. Switching to the new argument name resolves the issue and also fixes the deprecation messages.

More information: https://github.com/appuniversum/ember-appuniversum/issues/221