Remove all "default" tests (#58)
* Remove all "default" tests

These tests are failing since we can't render the field components like that. Removing them allows us to add some meaningful tests and enforcing that in CI.

* Don't allow the test step to fail

Since all failing tests are removed we can fail the pipeline when the test step fails. PRs that add failing tests won't be able to be merged.