ROLLBACK: simplification of adding of new entries
The previous version had a seperate creationForm, with intermediate
state to allow creating more complex objects than what is shown in the
listing.
E.g A Listing shows just the chapters titles of the book, but if you want
to add a chapter, in a separate create form, you could ask for adding
content too.

It reqires way more thought and implementationwork to allow something
that has a sane UX behaviour in different contexts.
Probably we will need to extend the form model for this.