Made sure we are reverse compatible for lang-strings
- see comments in code

Note: I added extra filter for literal.
This technically breaks reverse-compatibility here but, if we start
working on non-literal values for an input field, we will break data
for sure.
Most likely if non-literals are displayed, either something is wrong
with data or with the form configuration.

so I consider this acceptable.