Remove `@renderInPlace` from the ConceptSchemeSelector component (#86)
`ConceptSchemeSelector` is the only component that uses the `@renderInPlace` argument, which makes the component behave differently from the others selects. This difference in behavior is problematic for for the automated testing setup. Removing the argument makes the behavior consistent with the other select components.

`@renderInPlace` should only be needed if the wormhole based rendering causes problems, which can happen if the select is used in things like modals.