WIP show the table title again
This assumes that the Appuniversum table doesn't output any whitespace if no data is inserted into the title block. Ember doesn't support conditional named blocks, so we need to depend on hacks like this, otherwise we would have to duplicate the whole component simply to not pass a title block in some cases.