Update Ember Data to 4.12.0 (#127)
It seems we were using a wide version range for Ember Data and the last release made the Embroider builds fail if the built-in packageRules and adapters weren't disabled.

This change updates Ember Data to 4.12 and disables the compat adapters and packageRules. Once a new version of `@embroider/compat` we can remove the config again.