rename the `updateValue` task to `search`
The parent class defines `updateValue` as a normal method but the new task syntax doesn't allow calling it with `super` since it's no longer a method but a class property. Renaming the task allows us to call `this.updateValue` instead, which also is less confusing. It is also consistent with the other search components.