Make sure the remote url preview link is displayed
We accidentally introduced a regression when resolving Appuniversum deprecations. The 2-way-binding that was previously used also notified the rendering system so the link was displayed. When we switched to regular events for content updates we forgot to mark the address property as tracked and as a result the template wouldn't rerender anymore.