Fix some auto focus issues
Appuniversum v3.1 removed their internal use of the `@zestia/ember-auto-focus` addon. It seems we were implicitly depending on this addon as well. Instead of adding the addon to our own dependency list, I opted for a custom solution, since that means we have less dependencies to maintain, and it's a very simple implementation for our use-case.

This also fixes the issue where the input fields would be focused when the forms are rendered for the first time. Now it only focuses a field when the user adds a new entry.