Fix an issue with the "RemoteUrl" field
When the user creates a new link and then immediately clicks the delete button nothing happens. The user has to click the button a second time.

The reason for this is that we use the blur event on the input field to handle validations and this event is triggered _before_ the click event of the button. Because the field is empty error messages are displayed and the button is pushed down and because of that displacement the click event isn't triggered.

By switching to the `mousedown` event the removal happens before the blur event is triggered, which sidesteps the problem.

DL-5449